PyOs Sandbox a0.1
I'm happy to see you!
(this is first ever version of this program)

There isn't many features,
beacuse i wanted to make the core of this sandbox first, and not
make a ton of features that can't be used.

This is LINUX ONLY!!! This version COULD be stable on windows,
but it's not supported and never will be.